voice-call-bot/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pandas==2.1.3
openpyxl==3.1.2
twilio==8.10.0
openai==1.3.5
pydantic==2.5.0
python-multipart==0.0.6
apscheduler==3.10.4
httpx==0.25.1
